+++ packages/xsoulspace_foundation/lib/src/extensions/build_context_x.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 ///
 /// @ai Use this extension to easily retrieve theme-related information in your
 /// widgets without repetitive code.
-extension BuildContextX on BuildContext {
+extension XSBuildContextX on BuildContext {
   /// The current [ThemeData] for the context.
   ///
   /// @ai Use this property to access theme-related properties such as colors,
@@ -24,4 +24,29 @@ extension BuildContextX on BuildContext {
   /// @ai Use this property to access color definitions for the current theme,
   /// ensuring consistent color usage throughout your application.
   ColorScheme get colorScheme => theme.colorScheme;
+
+  /// Retrieves the view padding for the current context.
+  ///
+  /// This method calculates the padding by comparing the [View] padding
+  /// and the [MediaQuery] padding, returning the greater value for each side.
+  ///
+  /// @ai Use this method to obtain accurate padding values for your widgets,
+  /// especially when dealing with safe areas and insets.
+  EdgeInsets get viewPadding {
+    final view = View.of(this);
+    final vPad = EdgeInsets.only(
+      top: view.padding.top / view.devicePixelRatio,
+      bottom: view.padding.bottom / view.devicePixelRatio,
+      left: view.padding.left / view.devicePixelRatio,
+      right: view.padding.right / view.devicePixelRatio,
+    );
+    final cPad = MediaQuery.paddingOf(this);
+
+    return EdgeInsets.only(
+      top: vPad.top > cPad.top ? vPad.top : cPad.top,
+      bottom: vPad.bottom > cPad.bottom ? vPad.bottom : cPad.bottom,
+      left: vPad.left > cPad.left ? vPad.left : cPad.left,
+      right: vPad.right > cPad.right ? vPad.right : cPad.right,
+    );
+  }
 }