+++ packages/xsoulspace_review/lib/src/store_reviewer/reviewers/web_store_reviewer.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/widgets.dart';
+
+import '../store_reviewer.dart';
+
+final class WebStoreReviewer extends StoreReviewer {
+  @override
+  Future<void> requestReview(
+    final BuildContext context, {
+    final Locale? locale,
+    final bool force = false,
+  }) async {
+    // TODO(arenukvern): add review request for web
+  }
+}