+++ packages/xsoulspace_review/lib/xsoulspace_review.dart
@@ -1,0 +1,4 @@
+export 'package:wiredash/wiredash.dart';
+
+export 'src/store_reviewer/store_reviewer.dart';
+export 'src/user_feedback/user_feedback.dart';